# src/deployment/swaggerSchemaAPIGateway.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Vianaar",
    "description": "",
    "version": "1.0.0",
    "contact": {
      "name": "",
      "url": "",
      "email": ""
    }
  },
  "components": {
    "securitySchemes": {},
...
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/units/{unit_id}/catamaran": {
      "get": {
        "tags": ["Service"],
        "description": "API to fetch all catamaran request",
        "parameters": [
          {
            "schema": {
              "type": "number",
              "default": 10
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "number",
              "default": 0
            },
            "in": "query",
            "name": "offset",
            "required": false
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "project_id",
            "required": true
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "unit_id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "book_in_date": {
                        "type": "string",
                        "format": "date"
                      },
                      "cost": {
                        "type": "number"
                      },
                      "from_time": {
                        "type": "string",
                        "format": "time"
                      },
                      "to_time": {
                        "type": "string",
                        "format": "time"
                      },
                      "comments": {
                        "type": "string"
                      },
                      "admin_comments": {
                        "type": ["null", "string"]
                      },
                      "request_status": {
                        "type": "string"
                      },
                      "payment_status": {
                        "type": "string"
                      },
                      "add_time": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "responses": {
            "default": {
              "statusCode": "200"
            }
          },
          "uri": "http://vianaar-backend.eba-hxvhbm2d.ap-south-1.elasticbeanstalk.com/api/v1/projects/{project_id}/units/{unit_id}/catamaran",
          "passthroughBehavior": "when_no_match",
          "httpMethod": "get",
          "type": "http",
          "integrationHttpMethod": "get",
          "requestParameters": {
            "integration.request.path.project_id": "method.request.path.project_id",
            "integration.request.path.unit_id": "method.request.path.unit_id",
            "integration.request.header.Content-Type": "'application/json'"
          }
        }
      },
      "post": {
        "tags": ["Service"],
        "description": "API to create catamaran request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "book_in_date": {
                    "type": "string",
                    "format": "date"
                  },
                  "slot_id": {
                    "type": "number"
                  },
                  "number_of_pax": {
                    "type": "number",
                    "minimum": 1
                  }
                },
                "required": ["book_in_date", "slot_id", "number_of_pax"]
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "project_id",
            "required": true
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "unit_id",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "number"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "responses": {
            "default": {
              "statusCode": "201"
            }
          },
          "uri": "http://vianaar-backend.eba-hxvhbm2d.ap-south-1.elasticbeanstalk.com/api/v1/projects/{project_id}/units/{unit_id}/catamaran",
          "passthroughBehavior": "when_no_match",
          "httpMethod": "post",
          "type": "http",
          "integrationHttpMethod": "post",
          "requestParameters": {
            "integration.request.path.project_id": "method.request.path.project_id",
            "integration.request.path.unit_id": "method.request.path.unit_id",
            "integration.request.header.Content-Type": "'application/json'"
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/units/{unit_id}/catamaran/{request_id}": {
      "delete": {
        "tags": ["Service"],
        "description": "API to delete catamaran request for given request ID",
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "project_id",
            "required": true
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "unit_id",
            "required": true
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "request_id",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "responses": {
            "default": {
              "statusCode": "204"
            }
          },
          "uri": "http://vianaar-backend.eba-hxvhbm2d.ap-south-1.elasticbeanstalk.com/api/v1/projects/{project_id}/units/{unit_id}/catamaran/{request_id}",
          "passthroughBehavior": "when_no_match",
          "httpMethod": "delete",
          "type": "http",
          "integrationHttpMethod": "delete",
          "requestParameters": {
            "integration.request.path.project_id": "method.request.path.project_id",
            "integration.request.path.unit_id": "method.request.path.unit_id",
            "integration.request.path.request_id": "method.request.path.request_id",
            "integration.request.header.Content-Type": "'application/json'"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}